Prototype Saving (See extended description)
Files can be saved without polygon map. It will currently overwrite the
file in the location it was loaded from. Right now this accomplishes
nothing (obviously) but next step is to save polygon map over the top.
For testing the code for saving is tied to pressing the "s" key when the
labelbox is open.